requirements.txt
tensorflow
matplotlib
seaborn
numpy
pandas
PyYAML
argparse

## local package installation
-e .   ### to install local packages defined in setup.py
